requirements.txt
mysql-connector-python==8.0.20
irc==18.0.0